backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
websockets==15.0.1